+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 
 - [#3189](https://github.com/ChainSafe/forest/issues/3189): Changed the database
   organisation to use multiple columns. The database will need to be recreated.
+- [#3220](https://github.com/ChainSafe/forest/pull/3220): Removed the
+  `forest-cli chain validate-tipset-checkpoints` and
+  `forest-cli chain tipset-hash` commands.
 
 ### Added